New xmlutils package
