+++ internal/tags/tag_replacement.go
@@ -4,10 +4,18 @@ import (
 	"bytes"
 	"fmt"
 	"text/template"
+
+	"github.com/tomwatkins1994/go-docx-template/internal/xmlutils"
 )
 
-func ReplaceTagsInText(text string, data map[string]any, funcMap template.FuncMap) (string, error) {
-	tmpl, err := template.New("").Funcs(funcMap).Parse(text)
+func ReplaceTagsInText(xmlString string, data map[string]any, funcMap template.FuncMap) (string, error) {
+	// Prepare the XML for tag replacement
+	preparedXmlString, err := xmlutils.PrepareXmlForTagReplacement(xmlString)
+	if err != nil {
+		return "", err
+	}
+
+	tmpl, err := template.New("").Funcs(funcMap).Parse(preparedXmlString)
 	if err != nil {
 		return "", fmt.Errorf("error parsing template: %v", err)
 	}
@@ -18,5 +26,8 @@ func ReplaceTagsInText(text string, data map[string]any, funcMap template.FuncMa
 		return "", err
 	}
 
-	return buf.String(), err
+	// Fix any issues in the XML
+	outputXmlString := xmlutils.FixXmlIssuesPostTagReplacement(buf.String())
+
+	return outputXmlString, err
 }
