+++ internal/xmlutils/xml_manipulation_test.go
@@ -1,4 +1,4 @@
-package docxtpl
+package xmlutils
 
 import (
 	"testing"
@@ -75,7 +75,7 @@ func TestFixXmlIssuesPostTagReplacement(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			outputXml := fixXmlIssuesPostTagReplacement(tt.inputXml)
+			outputXml := FixXmlIssuesPostTagReplacement(tt.inputXml)
 			assert.Equal(t, outputXml, tt.expectedOutputXml)
 		})
 	}
